feat: modularize TrpcProvider
