+++ apps/web/src/components/BeneficiaryForm/BeneficiaryFormCreation.tsx
@@ -5,7 +5,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { Options } from '@mss/web/utils/options'
 import { trpc } from '@mss/web/trpc'
 import { useRouter } from 'next/navigation'
-import { withTrpc } from '@mss/web/withTrpc'
+import { withTrpc } from '@mss/web/components/TrpcProvider'
 import { Routes } from '@mss/web/app/routing/routes'
 import { MutationInput } from '@mss/web/features/createMutation.client'
 import { AddBeneficiaryWithGeneralInfoClient } from '@mss/web/features/beneficiary/addBeneficiary/addBeneficiaryWithGeneralInfo.client'
