+++ apps/web/src/components/DocumentsTab/AddDocumentModalForm.tsx
@@ -3,7 +3,7 @@
 import mime from 'mime-types'
 import { DefaultValues, useForm } from 'react-hook-form'
 import { trpc } from '@mss/web/trpc'
-import { withTrpc } from '@mss/web/withTrpc'
+import { withTrpc } from '@mss/web/components/TrpcProvider'
 import {
   SelectFormField,
   TagsFormField,
