+++ apps/web/src/components/FollowupForm/FollowupFormCreation.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { trpc } from '@mss/web/trpc'
 import { useRouter } from 'next/navigation'
-import { withTrpc } from '@mss/web/withTrpc'
+import { withTrpc } from '@mss/web/components/TrpcProvider'
 import { Routes } from '@mss/web/app/routing/routes'
 import { MutationInput } from '@mss/web/features/createMutation.client'
 import { AddFollowupClient } from '@mss/web/features/followup/addFollowup.client'
