+++ apps/web/src/components/StructureForm/StructureFormEdition.tsx
@@ -5,7 +5,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 
 import { trpc } from '@mss/web/trpc'
 import { useRouter } from 'next/navigation'
-import { withTrpc } from '@mss/web/withTrpc'
+import { withTrpc } from '@mss/web/components/TrpcProvider'
 import { deserialize, Serialized } from '@mss/web/utils/serialization'
 import { groupFollowupTypesByLegality } from '@mss/web/helper/groupFollowupTypes'
 import { useMemo, useState } from 'react'
