+++ apps/web/src/components/TrpcProvider/index.ts
@@ -1,0 +1,3 @@
+export { TrpcProvider } from './TrpcProvider'
+export { useTrpcClient } from './useTrpcClient'
+export { withTrpc } from './withTrpc'
