+++ apps/web/src/components/UserForm/UserFormEdition.tsx
@@ -4,7 +4,7 @@ import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { trpc } from '@mss/web/trpc'
 import { useRouter } from 'next/navigation'
-import { withTrpc } from '@mss/web/withTrpc'
+import { withTrpc } from '@mss/web/components/TrpcProvider'
 import { deserialize, Serialized } from '@mss/web/utils/serialization'
 import type { MutationServerState } from '@mss/web/features/createMutation.server'
 import type { MutationInput } from '@mss/web/features/createMutation.client'
